# morfinator_api/models/ingredient.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from morfinator_api.models.base_model_ import Model
from morfinator_api import util


class Ingredient(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: int=None, name: str=None, type: str=None):  # noqa: E501
        """Ingredient - a model defined in Swagger

        :param id: The id of this Ingredient.  # noqa: E501
        :type id: int
        :param name: The name of this Ingredient.  # noqa: E501
        :type name: str
        :param type: The type of this Ingredient.  # noqa: E501
        :type type: str
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'type': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'type': 'type'
        }

        self._id = id
        self._name = name
        self._type = type

    @classmethod
    def from_dict(cls, dikt) -> 'Ingredient':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Ingredient of this Ingredient.  # noqa: E501
        :rtype: Ingredient
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Ingredient.


        :return: The id of this Ingredient.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Ingredient.


        :param id: The id of this Ingredient.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Ingredient.


        :return: The name of this Ingredient.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Ingredient.


        :param name: The name of this Ingredient.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def type(self) -> str:
        """Gets the type of this Ingredient.


        :return: The type of this Ingredient.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Ingredient.


        :param type: The type of this Ingredient.
        :type type: str
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type