# morfinator_api/models/user_meals.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from morfinator_api.models.base_model_ import Model
from morfinator_api.models.meal import Meal
from morfinator_api import util


class UserMeals(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, user_id: int=None, meals: List[Meal]=None):  # noqa: E501
        """UserMeals - a model defined in Swagger

        :param user_id: The user_id of this UserMeals.  # noqa: E501
        :type user_id: int
        :param meals: The meals of this UserMeals.  # noqa: E501
        :type meals: List[Meal]
        """
        self.swagger_types = {
            'user_id': int,
            'meals': List[Meal]
        }

        self.attribute_map = {
            'user_id': 'user_id',
            'meals': 'meals'
        }

        self._user_id = user_id
        self._meals = meals

    @classmethod
    def from_dict(cls, dikt) -> 'UserMeals':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserMeals of this UserMeals.  # noqa: E501
        :rtype: UserMeals
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self) -> int:
        """Gets the user_id of this UserMeals.


        :return: The user_id of this UserMeals.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this UserMeals.


        :param user_id: The user_id of this UserMeals.
        :type user_id: int
        """

        self._user_id = user_id

    @property
    def meals(self) -> List[Meal]:
        """Gets the meals of this UserMeals.


        :return: The meals of this UserMeals.
        :rtype: List[Meal]
        """
        return self._meals

    @meals.setter
    def meals(self, meals: List[Meal]):
        """Sets the meals of this UserMeals.


        :param meals: The meals of this UserMeals.
        :type meals: List[Meal]
        """

        self._meals = meals